requirements.txt
azure-functions
azure-communication-email